storefront/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
mysqlclient==2.2.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.1.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
whitenoise==6.6.0


uWSGI==2.0.20